ajustes no modulo Unity, paginacao, dependencia e teste unitario

+++ SysManager.Application/Services/UnityService.cs
@@ -14,13 +14,15 @@ namespace SysManager.Application.Services
     public class UnityService 
     {
         private readonly UnityRepository _unityRepository;
+        private readonly ProductRepository _productRepository;
 
-        public UnityService(UnityRepository unityRepository)
+        public UnityService(UnityRepository unityRepository, ProductRepository productRepository)
         {
             this._unityRepository = unityRepository;
+            this._productRepository = productRepository;
         }
 
-        public async Task<ResultData> PostAsync(UnityPostRequest request)
+        public virtual async Task<ResultData> PostAsync(UnityPostRequest request)
         {
             var validator = new UnityPostRequestValidator(_unityRepository);
             var validatorResult = validator.Validate(request);
@@ -32,7 +34,7 @@ namespace SysManager.Application.Services
             return Utils.SuccessData(response);
         }
 
-        public async Task<ResultData> PutAsync(UnityPutRequest request)
+        public virtual async Task<ResultData> PutAsync(UnityPutRequest request)
         {
             var validator = new UnityPutRequestValidator(_unityRepository);
             var validatorResult = validator.Validate(request);
@@ -45,7 +47,7 @@ namespace SysManager.Application.Services
             return Utils.SuccessData(response);
         }
 
-        public async Task<ResultData> GetByIdAsync(Guid id)
+        public virtual async Task<ResultData> GetByIdAsync(Guid id)
         {
             var response = await _unityRepository.GetByIdAsync(id);
 
@@ -55,17 +57,25 @@ namespace SysManager.Application.Services
             return Utils.ErrorData(SysManagerErrors.Unity_Get_BadRequest_Id_Is_Invalid_Or_Inexistent.Description());
         }
 
-        public async Task<ResultData> GetByFilterAsync(UnityGetFilterRequest request)
+        public virtual async Task<ResultData> GetByFilterAsync(UnityGetFilterRequest request)
         {
             var result = await _unityRepository.GetByFilterAsync(request);
             return Utils.SuccessData(result);
         }
 
-        public async Task<ResultData> DeleteByIdAsync(Guid id)
+        public virtual async Task<ResultData> DeleteByIdAsync(Guid id)
         {
             var response = await _unityRepository.GetByIdAsync(id);
             if (response != null)
             {
+                var count = await _productRepository.GetCountProductByDependenceAsync("unityId", id);
+
+                var message = string.Format(SysManagerErrors.Unity_Delete_BadRequest_Id_Gave_Dependence.Description(), count);
+                if (count > 0)
+                {
+                    return Utils.ErrorData(message);
+                }
+
                 var result = await _unityRepository.DeleteByIdAsync(id);
                 return Utils.SuccessData(result);
             }
